requirements.txt
parse
pyyaml
boto3==1.3.1
awscli==1.10.67
